+++ src/main/java/com/programmers/cafe/exception/DataNotFoundException.java
@@ -1,4 +1,4 @@
-package com.programmers.cafe.service;
+package com.programmers.cafe.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
